bucket-crud/requirements.txt
google-cloud-storage>=2.18.0,<3.0.0
google-auth>=2.30.0,<3.0.0
google-auth-oauthlib>=1.2.0,<2.0.0

